+++ CHANGELOG.md
@@ -5,6 +5,10 @@
  - Fix PiOLED LCD from changing I2C address when options are saved ([#579](https://github.com/kizniche/mycodo/issues/579))
  - Fix Generic 16x2/16x4 LCD display issue ([#578](https://github.com/kizniche/mycodo/issues/578))
 
+### Features
+
+ - Add Input: SHT31 Smart Gadget Bluetooth humidity/temperature sensor ([#559](https://github.com/kizniche/mycodo/issues/559))
+
 
 ## 7.0.0 (2018-12-08)